refact test
